+++ CHANGELOG.md
@@ -17,6 +17,7 @@ contents to a new release section (e.g. _[0.x]_).
 - Added some project settings for when using vscode, instead of ignoring those files in
   git. (PLUT-598)
 - SDK: add latex rendering for greek-symbols and trigonometric functions (PLUT-714)
+- SDK: add latex rendering for natural-log, log base 10 & general log functions (PLUT-805)
 
 ### Fixed